rm bootstrap.txt and ignore it.
We will describe the setup in the readme and document the used software versions inside the ipython notebook. All other things will get outdated and we always tend to forget to update the bootstrap.txt. Also there are too many packages in there that are basically just deps of deps of deps.
Installing the basic numpy/scipy/pandas/mpl/ipython is what we need. Plus perhaps HDF tables and pyprind. This is less intimidating than to have such a large bootstrap file.